Fix `injectHints` might return undefined markers
The `inner` function of `injectHints` now returns an empty array instead
of undefined if no markers could be produced, since all results of the
`inner` function are recursively concatenated. VimFx seems to have worked
well even though some entries in the array of markers returned by
`injectHints` could be undefined. However, on the Huffman branch, the
undefined entries sometimes caused a TypeError to be thrown, which
eventually resulted in that none of the hint markers were given any hints.