Fix typo
Line 64 in marker.coffee contains a reference to `@markedElement`, though
that is the only occurance in the entire files. It clearly looks like it
is a typo of `@markerElement`, by inspecting the nearby code. The odd
thing, though, is that that line seems to have been added by commit
b5f488df, and already there the typo exists. Has the code ever worked?

I'm not sure if this fix currently fixes any issues, but it is needed in
the Huffman branch in order not to throw errors.