Use inputs instead of prompts for key customization
The help dialog looks exactly like before. When clicking a shortcut it is
no longer deleted. Instead the cursor is placed there, letting the user
edit it inline, which feels very natural. Clicking the + button no longer
opens a prompt. Instead it simply adds a new blank shortcut and focuses
it, ready for typing from the user. Shortcuts can be deleted by simply
editing one and backspacing all the chars.

Spaces are prevented from being typed, but can be pasted in. If so they
are stripped later on. Could possibly be improved via the paste event.

Enter blurs the inputs, which feels very natural.

NOTE: There are currently no checks for conflicting shortcuts. When the
prompts were used, at least the user was alerted if an identical key
already exists. That's not enough, though. This must be added later on.

Adding too many or too long shortcuts breaks the layout. It was a bit
buggy before too, though. Needs looking into.