Fix long shortcuts break layout (initial attempt)
The help dialog now only allow shortcuts that are 10 chars or less long,
in order not to break the layout. (Such long shortcuts are silly anyways.)

When the shortcuts cannot fit one line anymore, they now continue on the
next line.

Note: This is a little rough around the edges.