Fix: Esc command now works nicer in the help dialog
Before it was impossible to blur a key customization input using the Esc
command without also closing the help dialog. Also, if the user used the f
command to focus something in the help dialog, but changed their mind
half-way and wanted to remove the hints, that was also impossible to do
without also closing the help dialog. That was very annoying.

Now, basically, the help dialog is only removed by the Esc command if it
didn't do anything else (kind of).