Move `z-index` out of `.VimFxReset`
The hint markers need to set `z-index` dynamically (to rotate overlapping
markers). style.css is inserted as an agent sheet and uses `!important`
everywhere (this is needed to override userstyles, see commit 6145382),
and can therefore not be overrided by inline styles.

Therefore, I've moved `z-index` out of `.VimFxReset`, to each place that
actually _needs_ it: A few positioned elements.