Fix: `Vim::enterNormalMode` should reset `@keys`
Otherwise keys accumulated in other modes might be interpreted when going
back to normal mode. For example, following a link with a hint ending with
"a" and then directly trying to follow another link, would do so in a new
tab. That last "a" was still in `@keys` when returning to normal mode, so
when trying to use the "f" command, it was interpreted as "af" instead.