Fix: Dropped out of insert mode on location change
For example, it would be nice to usually be in insert mode in gmail, and
only occasionally leave it. However, as soon as you for example went from
the inbox to an e-mail, the location changed, which caused normal mode to
be entered. Only when in hints mode we need to do that (issue #101).