Let un-blacklist remove _all_ matching rules
Previously, only the _first_ matching rule was removed when the user
clicked the "un-blacklist button". But it would be a bummer if the user
clicks to remove the page from the blacklist, and it stays blacklisted.

Therefore _all_ matching rules are displayed and removed. To facilitate
this, the string in the textbox of the toolbar button is treated as a
blacklist of its own, which should be removed from or concatenated with
the blacklist pref. (Which means that the user is now able to add multiple
rules at the same time, should anyone want that.)