Fix #213: Suppress logic bug
For example, if you press 'x' to close the current tab, it will close
before keyup fires. So keyup (and perhaps keypress) will fire in another
tab. Even if that particular tab is blacklisted, we must suppress the
event, so that 'x' isn't sent to the page. The rule is simple: If the
`suppress` flag is `true`, the event should be suppressed, no matter what.
It has the highest priority.

Previously, `vim.blacklisted` checks were performed even in keypress and
keyup, which caused the suppress to be ignored.