Auto-focus findbar in find mode onInput
Now, if you unfocus the text input of the findbar, it will be
automatically focused again when you start typing. Before, nothing at all
happened (commands weren't run either), which felt really weird. This
reminds you that you're still in find mode, and need to press Enter or Esc
to exit from it.