Fix #280: Use `[rel]` for prev/next link matching
The standard way of marking up the previous or next page of a series is
to use a `<link>` or `<a>` element with a `rel` attribute.

We don’t bother with `<link>`s since the commands are substitutes for
clicking a visible element. (And at least in my experience, they’re not
commonly used.) Moreover, if there’s a `<link>` with a `rel` attribute,
there’s likely a corresponding `<a>`, too.

We first try to find a nicely marked up link, then we fall back on
pattern matching of the text contents of the links.

The spec on the `rel` attribute:
<http://www.w3.org/TR/html5/links.html#linkTypes>

How Google handles it:
<https://support.google.com/webmasters/answer/1663744?hl=en>

This commit is based on #313 by @zhuochun.