fixes issue #363
There is a bug/issue with Firefox where the value of 'general.useragent.locale'
is 'chrome://global/locale/intl.properties' and not a language code eg 'de-DE'.
This prevents us from guessing the correct path to the l10n file.

This commit is a nifty hack, which adds yet another string bundle to be used
for localization. That might lead to better localisation.