Remove bloomfilter
The bloomfilter can run out of space, and actually does so pretty quickly:
See #176. When it is full it is not of any help anymore. All it does then is
unconditionally giving better hints to `<a>` elements. But since commit 01657421
that is done in a better way.

In summary, the bloomfilter is not useful anymore.