Fix #60: Give the same hint to links with the same URL
This results in fewer hints, which results in shorter hints overall.

'vf' is excluded so that everything still is focusable.

`<a href="#">` links, which are commonly used as buttons, are excluded since
clicking them likely does different things via JavaScript.

`<a href="javascript:">` links are also excluded, mainly because
`<a href="javascript:void(0)">` is a commonly used as buttons as well.