Provide `vim.state`
Instead of just assigning random properties on `vim` objects (such as
`.blacklisted` and `.lastInteraction`) we now assign them to `vim.state`. This
way we can properly initialize those values when new `vim` objects are made, and
easily reset all of them when the URL changes (which means that the state
changes).