Do not consider links without href as proper links
Links ending with a `#` (`<a href="#">`) as well as `javascript:` links are
already excluded, because they're used as buttons, not links. Links without href
at all (`<a>`) should be excluded too, since they are also used as buttons (for
example, the voting buttons on stackoverflow.com). If they're not excluded, all
such links get the same hint, even though all of them do different things
through JavaScript. Moreover, the 'F' and 'gF' commands would try to open them
in new tabs, which of course wouldn't work.