Reduce the number of hints on Google Calendar
As noted in https://github.com/akhodakivskiy/VimFx/pull/451#issuecomment-71582664,
Google Calendar gets too many hints. That's because many buttons are implemented
as `<div>`s with lots of `<div>`s inside, all of which have "button" in the
their class names. This commit exludes elements with "button" in the class name
if they contain yet such an element. This works fine on Google Calendar.
Hopefully it does not exlude too much on other pages.