Drastically improve speed up hints generation
We used to apply a class to each element, to work around `border-radius`.
Removing that drastically speeds up hints generation. The "loss" is that the
markers no longer appear in the top-left corner, but rather in the left-middle.
That takes a little bit of getting used to, but is definitely worth it.