Rewrite the README key features / shortcuts parts
This commit merges those parts into one. The complete list of shortcuts has been
removed. As it will only grow, it gets a little overwhelming. It is also a PITA
to maintain. Instead, only some of the most important default shortcuts are
mentioned, more like on the current AMO page, and there's a link to
lib/defaults.coffee in case somebody wants to easily view all defaults before
installing the add-on.

The above makes the readme suitable as the text used on AMO. Therefore I also
updated a few links to make them work on AMO as well.