Simplify compatible Firefox version management
The only thing that matters is the minimum version, and it is only the major
number of it that matters.

Including a maximum version number is pretty pointless since most of the time
we've just set it to some version in the future without actually knowing if
VimFx works with that version or not. Firefox's release cycle is also so fast
that it is very hard to keep up with it. Finally, AMO automatically bumps the
maximum version if the addon passes their tests. So it is better to simply set
it to `*` and be done with it.