Improve the font size in the help dialog
When testing on a different computer I noticed the help dialog text being way
too large. I tried adjusting the CSS, but there was no sweet spot where the font
size was pleasant on both my regular computer and the other one I tried on.

This commits changes the default font size is the same as for menus, which is
usually quite small. It is then increased programmatically until just before a
scrollbar is caused.