Change `vf` into `zf` to free up `v`
Using the default shortcuts, `a`, `g`, `y` and `v` are prefix keys. `v` is only
used in `vf`. While `v` in `vf` kind of makes sense if you come from vim, it
would make a lot more sense to use `v` for a possible future text selection mode
(#480). `zf` is better since `z` is a prefix key in vim as well. The `z`
"namespace" will be useful as we add more commands.