Set prevent_autofocus to false by default
"VimFx will be nice to your browser and to your habits. Promise."

Turning it on by default have proven to be too much of a change. When people
become annoyed by autofocus they can enable it.

As far is I can tell, this is also how Vimium and Vimperator behaves.