Fix shortcut triggering on slowly loadnig pages
Frame scripts cannot receive events until a DOMWindow has been created in it. If
you opened a link in a new tab, and the page loaded very slowly making it take a
long time before a DOMWindow is created, it meant that no VimFx shortcuts could
used during that time.

The solution is to handle shortcuts in the main process until a DOMWindow has
been created.