Handle {,non-}pinned tabs separately in `gJ` and `gK`
Previously, it was possible to move a non-pinned tab into the pinned tabs,
making it pinned, and to move a pinnned tab into the non-pinned tabs, making it
non-pinned.

Now, that's not the case anymore. The first non-pinned tab wraps to the last
tab, and the last tab wraps to the first non-pinned tab, and vice versa for
pinned tabs.

In other words, you can no longer use `gJ` and `gK` to change the pinned state
of a tab. Only `gp` can.

See #606.